Refactor frontend to put helper apis in separate files
