+++ frontend/CMakeLists.txt
@@ -23,9 +23,13 @@ install(DIRECTORY
 
 install(PROGRAMS
   src/app.py
-  src/frontend_handler.py
-  src/log_helper.py
   DESTINATION lib/${PROJECT_NAME}
+)
+
+install(PROGRAMS
+  src/frontend_utils/frontend_handler.py
+  src/frontend_utils/log_helper.py
+  src/frontend_utils/recording_api.py
 )
 
 install(DIRECTORY
