+++ frontend/src/frontend_utils/frontend_handler.py
@@ -1,5 +1,4 @@
 from shared_msgs.msg import FinalThrustMsg
-import json
 
 def handle_frontend_event(node, event, data):
