+++ frontend/src/frontend_utils/log_helper.py
@@ -1,8 +1,6 @@
 import os
 import time
 import threading
-import json
-from datetime import datetime
 import re
 
 class LogHelper:
